move common transformTemplate
