+++ src/webpage.js
@@ -21,6 +21,7 @@ const md = require('markdown-it')({
     linkify: true,
     typographer: true,
 });
+const { transformTemplate } = require('./renderHTML');
 
 // use mdemoji plugin
 md.use(mdemoji);
@@ -28,37 +29,6 @@ md.use(mdemoji);
 md.renderer.rules.emoji = (token, idx) => `<i class="twa twa-${token[idx].markup}"></i>`;
 
 const defaultTemplatePath = 'src/template/web/index.html';
-
-
-/**
- * Using the given parameters, calls the Mustache engine
- * and renders the HTML page.
- * @param {string} templateFile Path for template file
- * @param {string} contractName Contract name
- * @param {object} contractData Contract data, containing ast and comments to be rendered
- * @param {string} contractPath Path to contract file
- */
-function transformTemplate(
-    templateFile, contractName, contractData, contractPath, contractsStructure, hasLICENSE,
-) {
-    // read template into a string
-    const templateContent = String(fs.readFileSync(templateFile));
-    // put all data together
-    const view = {
-        filePath: contractPath,
-        contract: {
-            name: contractName,
-        },
-        contractData,
-        contractsStructure,
-        currentDate: new Date(),
-        hasLICENSE,
-        CONTRACT: true,
-    };
-    // calls the render engine
-    const output = Mustache.render(templateContent, view);
-    return output;
-}
 
 /**
  * To write!
